+++ src/app/modules/controls-node/reducer/index.ts
@@ -1,0 +1,1 @@
+export * from './node-controls.reducer';
