+++ src/app/modules/controls-node/reducer/node-controls.reducer.ts
@@ -1,0 +1,46 @@
+import { createReducer, on } from '@ngrx/store';
+import * as nodeControlActions from '../actions/node-controls.actions';
+
+export interface NodeControlState {
+    renderNodes: boolean;
+    nodeSize: number;
+    minNodeSize: number;
+    uniformNodeSize: boolean;
+    pointAttenuation: boolean;
+    pointUseIcon: boolean;
+}
+
+const initialState: NodeControlState = {
+    renderNodes: true,
+    uniformNodeSize: false,
+    nodeSize: 100,
+    minNodeSize: 0.1,
+    pointAttenuation: true,
+    pointUseIcon: true,
+};
+
+export const reducer = createReducer(
+    initialState,
+    on(nodeControlActions.renderNodes, (state, { value }) => ({ ...state, renderNodes: value })),
+    on(nodeControlActions.setNodeSize, (state, { nodeSize }) => ({
+        ...state,
+        nodeSize: nodeSize,
+        minNodeSize: state.minNodeSize > nodeSize ? nodeSize : state.minNodeSize,
+    })),
+    on(nodeControlActions.setMinimumNodeSize, (state, { nodeSize }) => ({
+        ...state,
+        minNodeSize: nodeSize,
+    })),
+    on(nodeControlActions.setPointAttenuation, (state, { value }) => ({
+        ...state,
+        pointAttenuation: value,
+    })),
+    on(nodeControlActions.setPointUseIcon, (state, { value }) => ({
+        ...state,
+        pointUseIcon: value,
+    })),
+    on(nodeControlActions.setUniformNodeSize, (state, { value }) => ({
+        ...state,
+        uniformNodeSize: value,
+    })),
+);
