+++ src/app/modules/controls/components/channel-settings/channel-settings.component.ts
@@ -2,12 +2,12 @@ import { Component, OnInit } from '@angular/core';
 import { MatCheckboxChange } from '@angular/material/checkbox';
 import { Store } from '@ngrx/store';
 import { renderEdges, setEdgeUseDepthTest, setEdgeUseDottedLine } from '../../actions';
-import { ControlsState } from 'src/app/modules/controls/reducers/controls.reducer';
 import {
     selectEdgeDepthTest,
     selectEdgeDottedLine,
     shouldRenderEdges,
 } from 'src/app/modules/controls/selectors/controls.selectors';
+import { GenericControlsState } from '../../reducers';
 
 @Component({
     selector: 'app-channel-settings',
@@ -15,7 +15,7 @@ import {
     styleUrls: ['./channel-settings.component.scss'],
 })
 export class ChannelSettingsComponent {
-    constructor(private store: Store<ControlsState>) {}
+    constructor(private store: Store<GenericControlsState>) {}
 
     public shouldRenderEdges$ = this.store.select(shouldRenderEdges);
     public selectEdgeDepthTest$ = this.store.select(selectEdgeDepthTest);
