+++ src/app/modules/controls/components/misc-settings/misc-settings.component.ts
@@ -1,7 +1,7 @@
 import { Component } from '@angular/core';
 import { Store } from '@ngrx/store';
-import { ControlsState } from 'src/app/modules/controls/reducers/controls.reducer';
 import { resetControlsToDefault } from '../../actions';
+import { GenericControlsState } from '../../reducers';
 
 @Component({
     selector: 'app-misc-settings',
@@ -9,7 +9,7 @@ import { resetControlsToDefault } from '../../actions';
     styleUrls: ['./misc-settings.component.scss'],
 })
 export class MiscSettingsComponent {
-    constructor(private store: Store<ControlsState>) {}
+    constructor(private store: Store<GenericControlsState>) {}
 
     resetSettingsToDefault() {
         this.store.dispatch(resetControlsToDefault());
