+++ src/app/modules/controls/reducers/controls.reducer.ts
@@ -7,17 +7,11 @@ export const enum CameraFocusMode {
     Goto,
 }
 
-export interface ControlsState {
+export interface GenericControlsState {
     searchText: string;
-    renderNodes: boolean;
     renderEdges: boolean;
     renderLabels: boolean;
     minimumEdges: number;
-    nodeSize: number;
-    minNodeSize: number;
-    uniformNodeSize: boolean;
-    pointAttenuation: boolean;
-    pointUseIcon: boolean;
     cameraFov: number;
     cameraFocusMode: CameraFocusMode;
     edgeDepthTest: boolean;
@@ -26,17 +20,11 @@ export interface ControlsState {
     capacityFilterAmount: number;
 }
 
-const initialState: ControlsState = {
+const initialState: GenericControlsState = {
     searchText: '',
-    renderNodes: true,
     renderEdges: true,
     renderLabels: false,
-    uniformNodeSize: false,
     minimumEdges: 0,
-    nodeSize: 100,
-    minNodeSize: 0.1,
-    pointAttenuation: true,
-    pointUseIcon: true,
     cameraFov: 60,
     cameraFocusMode: CameraFocusMode.FocusOnly,
     edgeDepthTest: true,
@@ -50,27 +38,12 @@ export const reducer = createReducer(
     on(controlsActions.setSavedStateFromLocalStorage, (_state, { savedState }) => savedState),
     on(controlsActions.resetControlsToDefault, () => initialState),
     on(controlsActions.searchGraph, (state, { searchText }) => ({ ...state, searchText })),
-    on(controlsActions.renderNodes, (state, { value }) => ({ ...state, renderNodes: value })),
     on(controlsActions.renderEdges, (state, { value }) => ({ ...state, renderEdges: value })),
     on(controlsActions.renderLabels, (state, { value }) => ({ ...state, renderLabels: value })),
     on(controlsActions.minEdgesRecompute, (state, { minEdges }) => ({
         ...state,
         minimumEdges: minEdges,
     })),
-    on(controlsActions.setNodeSize, (state, { nodeSize }) => ({
-        ...state,
-        nodeSize: nodeSize,
-        minNodeSize: state.minNodeSize > nodeSize ? nodeSize : state.minNodeSize,
-    })),
-    on(controlsActions.setMinimumNodeSize, (state, { nodeSize }) => ({
-        ...state,
-        minNodeSize: nodeSize,
-    })),
-    on(controlsActions.setPointAttenuation, (state, { value }) => ({
-        ...state,
-        pointAttenuation: value,
-    })),
-    on(controlsActions.setPointUseIcon, (state, { value }) => ({ ...state, pointUseIcon: value })),
     on(controlsActions.setCameraFov, (state, { value }) => ({ ...state, cameraFov: value })),
     on(controlsActions.setCameraFocusMode, (state, { value }) => ({
         ...state,
@@ -91,9 +64,5 @@ export const reducer = createReducer(
     on(controlsActions.capacityFilterAmount, (state, { value }) => ({
         ...state,
         capacityFilterAmount: value,
-    })),
-    on(controlsActions.setUniformNodeSize, (state, { value }) => ({
-        ...state,
-        uniformNodeSize: value,
     })),
 );
